// src/App/services/graphql/queries/__generated__/get-pokemon-with-name-query.ts
/* eslint-disable */
/* This is an autogenerated file. Do not edit this file directly! */
import { TypedQueryDocumentNode } from "graphql";
export type getPokemonWithNameQuery = {
    gen3_species: ({
        name: string;
        id: number;
        pokemon_v2_pokemons: ({
            weight: number | null;
            height: number | null;
            pokemon_v2_pokemonabilities: ({
                pokemon_v2_ability: ({
                    name: string;
                }) | null;
            })[];
            pokemon_v2_pokemonstats: ({
                pokemon_v2_stat: ({
                    name: string;
                }) | null;
                base_stat: number;
            })[];
            pokemon_v2_pokemonsprites: ({
                sprites: string;
            })[];
        })[];
    })[];
};
export type getPokemonWithNameQueryVariables = {
    name: string;
};
export type getPokemonWithNameQueryDocument = TypedQueryDocumentNode<getPokemonWithNameQuery, getPokemonWithNameQueryVariables>;